requirements.txt
streamlit
pillow
torch
transformers
opencv-python
nltk
spacy
